+++ creusot/Cargo.toml
@@ -34,8 +34,13 @@ arraydeque = "0.4"
 creusot-contracts = { path = "../creusot-contracts", features = ["typechecker"] }
 escargot = { version = "0.5" }
 creusot-dev-config = { path = "../creusot-dev-config" }
+
 [[test]]
 name = "ui"
+harness = false
+
+[[test]]
+name = "creusot-contracts"
 harness = false
 
 [package.metadata.rust-analyzer]
