+++ creusot/tests/creusot-contracts/creusot-contracts.rs
@@ -1,0 +1,1 @@
+// dummy file, read during the `why3` test
